+++ src/include/tabulation.h
@@ -21,7 +21,7 @@
 namespace Smash {
 
 /**
- * A class for storing a one-dimensional lookup table of floating-point values.
+ * A class for storing a one-dimensional lookup table of doubleing-point values.
  */
 class Tabulation {
  public:
@@ -31,23 +31,23 @@ class Tabulation {
    * \param num number of intervals (the number of tabulated points is actually num+1)
    * \param f one-dimensional function f(x) which is supposed to be tabulated
    */
-  Tabulation(float x_min, float range, int num,
-             std::function<float(float)> f);
+  Tabulation(double x_min, double range, int num,
+             std::function<double(double)> f);
   /** Look up a value from the tabulation (without any interpolation, simply
    * using the closest tabulated value). If x is below the lower tabulation
    * bound we return 0, if it is above the upper bound we return the tabulated
    * value at the upper bound. */
-  float get_value_step(float x) const;
+  double get_value_step(double x) const;
   /** Look up a value from the tabulation using linear interpolation.
    * If x is above the upper bound, we use linear extrapolation of the two
    * highest tabulated points. */
-  float get_value_linear(float x) const;
+  double get_value_linear(double x) const;
 
  protected:
   // vector for storing tabulated values
-  std::vector<float> values_;
+  std::vector<double> values_;
   // lower bound and inverse step size 1/dx for tabulation
-  const float x_min_, inv_dx_;
+  const double x_min_, inv_dx_;
 };
 
 /**
@@ -63,8 +63,8 @@ class Tabulation {
  * \param[in] stable_mass mass of the stable particle in the final state
  * \param[in] type type of the resonance
  */
-inline float spec_func_integrand_1res(float resonance_mass, float sqrts,
-                               float stable_mass, const ParticleType &type) {
+inline double spec_func_integrand_1res(double resonance_mass, double sqrts,
+                               double stable_mass, const ParticleType &type) {
   if (sqrts <= stable_mass + resonance_mass) {
     return 0.;
   }
@@ -90,8 +90,8 @@ inline float spec_func_integrand_1res(float resonance_mass, float sqrts,
  * \param[in] t1 Type of the first resonance.
  * \param[in] t2 Type of the second resonance.
  */
-inline float spec_func_integrand_2res(float sqrts,
-                              float res_mass_1, float res_mass_2,
+inline double spec_func_integrand_2res(double sqrts,
+                              double res_mass_1, double res_mass_2,
                               const ParticleType &t1, const ParticleType &t2) {
   if (sqrts <= res_mass_1 + res_mass_2) {
     return 0.;
@@ -112,13 +112,13 @@ inline std::unique_ptr<Tabulation> spectral_integral_semistable(
     Integrator& integrate,
     const ParticleType& resonance,
     const ParticleType& stable,
-    float range) {
+    double range) {
   return make_unique<Tabulation>(
           resonance.min_mass_kinematic() + stable.mass(), range, 100,
-          [&](float srts) {
+          [&](double srts) {
             return integrate(resonance.min_mass_kinematic(),
                              srts - stable.mass(),
-                             [&](float m) {
+                             [&](double m) {
                                return spec_func_integrand_1res(m, srts,
                                                      stable.mass(), resonance);
                              });
@@ -130,15 +130,15 @@ inline std::unique_ptr<Tabulation> spectral_integral_unstable(
     Integrator2d& integrate2d,
     const ParticleType& res1,
     const ParticleType& res2,
-    float range) {
+    double range) {
     return make_unique<Tabulation>(
           res1.min_mass_kinematic() + res2.min_mass_kinematic(), range, 100,
-          [&](float srts) {
+          [&](double srts) {
             return integrate2d(res1.min_mass_kinematic(),
                                srts - res2.min_mass_kinematic(),
                                res2.min_mass_kinematic(),
                                srts - res1.min_mass_kinematic(),
-                               [&](float m1, float m2) {
+                               [&](double m1, double m2) {
                                  return spec_func_integrand_2res(srts, m1, m2,
                                                             res1, res2);
                                });
